ysl: Remove 'encryptable' flag from fstab file
On A13 devices, the Full Disk Encryption (FDE) feature is no longer supported. As a result, the 'encryptable' flag can only be used for adoptable storage, and not for full disk encryption.
This commit removes the 'encryptable' flag from the fstab file on A13 devices to prevent any issues with the system attempting to use FDE. Removing this flag will ensure that the system does not try to use FDE, and will instead use the adoptable storage feature as intended.
Note: This change only affects A13 devices and does not modify the fstab file on other devices.

Change-Id: Id38bcc4285b2e19931285a0269e3442a04817859
Signed-off-by: Kai Jones <kvjones0243@proton.me>
Signed-off-by: Rahmad Adi Pratama <OhMyVenyx@proton.me>